requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
emoji==2.8.0
fastapi==0.101.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
twilio==8.5.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
